requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
betterproto==2.0.0b4
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
frozenlist==1.3.3
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
steamio==0.9.6
tomli==2.0.1
typing_extensions==4.2.0
vdf==3.4
yarl==1.8.2
